+++ src/main/java/net/forcemaster_rpg/item/weapons/KnuckleItem.java
@@ -1,46 +1,21 @@
 package net.forcemaster_rpg.item.weapons;
 
-import net.forcemaster_rpg.effect.Effects;
-import net.minecraft.entity.EquipmentSlot;
-import net.minecraft.entity.LivingEntity;
-import net.minecraft.entity.effect.StatusEffectInstance;
-import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.item.ItemStack;
 import net.minecraft.item.ToolMaterial;
-import net.minecraft.util.math.Vec3d;
-import net.more_rpg_classes.effect.MRPGCEffects;
+import net.minecraft.item.tooltip.TooltipType;
+import net.minecraft.text.Text;
+import net.minecraft.util.Formatting;
 import net.spell_engine.api.item.weapon.SpellWeaponItem;
+
+import java.util.List;
 
 public class KnuckleItem extends SpellWeaponItem {
     public KnuckleItem(ToolMaterial material, Settings settings) {
         super(material, settings);
     }
 
-    @Override
-    public boolean postHit(ItemStack stack, LivingEntity target, LivingEntity attacker) {
-        int knuckle_chance_knockup = 15;
-        int randomrange_knockup = (int) ((Math.random() * (1 + knuckle_chance_knockup)) + 1);
-        int stonhand_stun_chance = 8;
-        int randomrange_stun = (int) ((Math.random() * (1 + stonhand_stun_chance)) + 1);
-        int stun_duration = 30;
-        PlayerEntity player = (PlayerEntity) attacker;
-
-        if (randomrange_knockup >= knuckle_chance_knockup ){
-        if (target.isPlayer()) {
-            Vec3d currentMovement = target.getVelocity();
-            target.setVelocity(currentMovement.x, currentMovement.y + 0.95, currentMovement.z);
-            target.velocityModified = true;
-        } else {
-            target.addVelocity(0, 0.4, 0);
-        }
-            stack.damage(1, attacker, EquipmentSlot.MAINHAND);
-    }
-        if(player.hasStatusEffect(Effects.STONE_HAND.registryEntry)){
-            if (randomrange_stun >= stonhand_stun_chance ){
-                target.addStatusEffect(new StatusEffectInstance(MRPGCEffects.STUNNED.registryEntry, stun_duration));
-            }
-        }
-        stack.damage(1, attacker, EquipmentSlot.MAINHAND);
-        return true;
+    public void appendTooltip(ItemStack stack, List<Text> tooltip, TooltipContext context, TooltipType type) {
+        tooltip.add(Text.translatable("item.forcemaster_rpg.knuckle.description_1").formatted(Formatting.RED));
+        super.appendTooltip(stack, context, tooltip, type);
     }
 }